updating bulk delete work

+++ resources/views/Admin/partial/livewire/ClearDeleteButtons.blade.php
@@ -3,8 +3,10 @@
             Clear  
             <i class="fa fa-search"></i>
         </button>
-        <button type="button" wire:click="deleteSelected('{{ $modelName }}')" class="btn btn-danger SearchButton">
-            Delete Selected
-            <i class="fa fa-trash"></i>
-        </button>                    
+        @if($showDeleteButton == 'true')
+            <button type="button" wire:click="deleteSelected('{{ $modelName }}')" class="btn btn-danger SearchButton">
+                Delete Selected
+                <i class="fa fa-trash"></i>
+            </button>                    
+        @endif 
 </div>
